% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{add_cortisol}
\alias{add_cortisol}
\title{Add Cortisol curve to existing E4 dataframe}
\usage{
add_cortisol(data)
}
\arguments{
\item{data}{The dataframe containing EDA signal}
}
\value{
Data frame with cortisol feature added
}
\description{
Add Cortisol curve to existing E4 dataframe
}
