% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{connect_peaks}
\alias{connect_peaks}
\title{Connects EDA peaks}
\usage{
connect_peaks(peaks, peak_height, window_size)
}
\arguments{
\item{peaks}{The EDA vector}

\item{peak_height}{EDA cut-off height to consider as a peak}

\item{window_size}{Window size of EDA peak}
}
\value{
The adjusted vector
}
\description{
Connects EDA peaks
}
