% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{downsample}
\alias{downsample}
\title{Downsample a series to match size of another series given by rate}
\usage{
downsample(data, rate)
}
\arguments{
\item{data}{A vector of sensor data}

\item{rate}{The sampling rate to downsize to}
}
\value{
A downsampled vector
}
\description{
Downsample a series to match size of another series given by rate
}
