% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_drive_data}
\alias{make_drive_data}
\title{Loads AffectiveROAD E4 Data Set}
\usage{
make_drive_data(folder)
}
\arguments{
\item{folder}{Folder containing the E4 data}
}
\value{
Data frame of AffectiveROAD data set
}
\description{
Loads AffectiveROAD E4 Data Set
}
