% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_hypertension_data}
\alias{make_hypertension_data}
\title{Loads Hypertension E4 Data Set}
\usage{
make_hypertension_data(
  folder,
  log_transform = FALSE,
  feature_engineering = FALSE
)
}
\arguments{
\item{folder}{Folder containing the E4 data}

\item{log_transform}{If TRUE, log-transforms EDA and HR signal}

\item{feature_engineering}{If TRUE, generates rolling features}
}
\value{
Data frame of HYPERTENSION data set
}
\description{
Loads Hypertension E4 Data Set
}
