% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_mmash_data}
\alias{make_mmash_data}
\title{Loads MMASH E4 Data Set}
\usage{
make_mmash_data(folder)
}
\arguments{
\item{folder}{Folder containing the E4 data}
}
\value{
Data frame of MMASH data set
}
\description{
Loads MMASH E4 Data Set
}
