% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_swell_data}
\alias{make_swell_data}
\title{Loads SWELL E4 Data Set}
\usage{
make_swell_data(folder, log_transform = FALSE, feature_engineering = FALSE)
}
\arguments{
\item{folder}{Folder containing the E4 data}

\item{log_transform}{If TRUE, log-transforms EDA and HR signal}

\item{feature_engineering}{If TRUE, generates rolling features}
}
\value{
Data frame of SWELL data set
}
\description{
Loads SWELL E4 Data Set
}
