% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_toadstool_data}
\alias{make_toadstool_data}
\title{Loads Toadstool E4 Data Set}
\usage{
make_toadstool_data(folder)
}
\arguments{
\item{folder}{Folder containing the E4 data}
}
\value{
Data frame of Toadstool data set
}
\description{
Loads Toadstool E4 Data Set
}
