% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{make_ubfc_data}
\alias{make_ubfc_data}
\title{Loads UBFC Data Set}
\usage{
make_ubfc_data(folder, log_transform = FALSE, feature_engineering = FALSE)
}
\arguments{
\item{folder}{Folder containing the data}

\item{log_transform}{If TRUE, log-transforms EDA and HR signal}

\item{feature_engineering}{If TRUE, generates rolling features}
}
\value{
Data frame of UBFC data set
}
\description{
Loads UBFC Data Set
}
