% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{model_cortisol}
\alias{model_cortisol}
\title{Model a Cortisol curve}
\usage{
model_cortisol(x, peak_height)
}
\arguments{
\item{x}{The EDA vector}

\item{peak_height}{EDA cut-off height to consider as a peak}
}
\value{
A vector of Cortisol measurements
}
\description{
Model a Cortisol curve
}
