% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{predict_kmeans}
\alias{predict_kmeans}
\title{Predict clusters on new data given existing kmeans model}
\usage{
predict_kmeans(object, newdata, method = c("centers", "classes"))
}
\arguments{
\item{object}{The kmeans model}

\item{newdata}{The new data frame}

\item{method}{Centers or Classes}
}
\value{
A vector of predictions
}
\description{
Predict clusters on new data given existing kmeans model
}
