% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{rdsamp}
\alias{rdsamp}
\title{Reads a a PhysioNET file}
\usage{
rdsamp(header.filename)
}
\arguments{
\item{header.filename}{The header file}
}
\value{
A data frame containing the data
}
\description{
Reads a a PhysioNET file
}
