% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{read.empatica}
\alias{read.empatica}
\title{Reads an Empatica E4 folder of data into a data frame}
\usage{
read.empatica(exdir, header.only = FALSE)
}
\arguments{
\item{exdir}{The folder containing the E4 data}

\item{header.only}{Header only, T or F}
}
\value{
A data frame containing the data
}
\description{
Reads an Empatica E4 folder of data into a data frame
}
