% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{rolling_features}
\alias{rolling_features}
\title{Creates new features from Empatica E4 data}
\usage{
rolling_features(data, window_size)
}
\arguments{
\item{data}{A data frame of EDA and HR sensor data}

\item{window_size}{Window size to use for rolling windows}
}
\value{
An expanded data frame of features
}
\description{
Creates new features from Empatica E4 data
}
