% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{rolling_features_simple}
\alias{rolling_features_simple}
\title{Creates new features from Empatica E4 data (HR only)}
\usage{
rolling_features_simple(data, window_size)
}
\arguments{
\item{data}{A data frame of HR sensor data}

\item{window_size}{Window size to use for rolling windows}
}
\value{
An expanded data frame of features
}
\description{
Creates new features from Empatica E4 data (HR only)
}
