% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{split_subject_data}
\alias{split_subject_data}
\title{Split a data frame by subject and class label}
\usage{
split_subject_data(data)
}
\arguments{
\item{data}{The data frame to split}
}
\value{
Split data set by subject
}
\description{
Split a data frame by subject and class label
}
