% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresshelpers.R
\name{wss}
\alias{wss}
\title{Compute Total Within-Cluster Sum of Square}
\usage{
wss(data, k)
}
\arguments{
\item{data}{The data frame to apply kmeans to}

\item{k}{The starting number of clusters}
}
\value{
The total within-cluster sum of sqaure
}
\description{
Compute Total Within-Cluster Sum of Square
}
